DDD - N-Layerd - Architecture. Working on Domain, respositories, services, and unit tests.

+++ src/Frameworks/TodoList.Application/TodoList/Commands/Contracts/ITodoListCommandsService.cs
@@ -5,6 +5,6 @@
         void AddItem(int id, string title, string description, string category);
         void UpdateItem(int id, string description);
         void RemoveItem(int id);
-        void RegisterProgression(int id, DateTime dateTime, decimal percent);
+        void RegisterProgression(int id, DateTime dateTime, int percent);
     }
 }
